refactor: 개선요청 상세 페이지 리팩터링
