+++ src/_hooks/fetcher/customer/useGetFeedbackInfoData.ts
@@ -3,13 +3,12 @@ import { useQuery } from "@tanstack/react-query";
 
 interface useGetFeedbackInfoDataProps {
   id: number;
-  token: string;
 }
 
-const useGetFeedbackInfoData = ({id, token}: useGetFeedbackInfoDataProps) => {
+const useGetFeedbackInfoData = ({id}: useGetFeedbackInfoDataProps) => {
   return useQuery({
     queryKey: ["feedbackInfo", id],
-    queryFn: () => getFeedbackInfoData({id, token}),
+    queryFn: () => getFeedbackInfoData({id}),
     retry: 1,
     staleTime: 1000 * 60 * 5,
     gcTime: 1000 * 60 * 30,
